Refactor stuffs
